+++ src/client/app/parser.js
@@ -45,13 +45,6 @@ var Parser = class Parser {
             // Set default widget type
             props.type =  tab ? 'tab' : props.type || 'fader'
 
-            // Backward compatibility patches
-            if (props.path) props.address = props.path
-            if (props.noPip) props.pips = !props.noPip
-            if (props.noSync) {
-                props.bypass = props.noSync
-                delete props.noSync
-            }
             // Safe copy widget's options
             let defaults = this.widgets[props.type].defaults()._props()